requirements.txt
google-genai>=1.0.0
google-adk>=1.0.0
aiohttp>=3.9.0
requests>=2.31.0
litellm>=1.0.0
python-dotenv>=1.0.0
mcp>=1.0.0
httpx>=0.27.0
pypdf>=4.0.0
